Carga da caixa de seleção de tipo de unidade comercial.

+++ Programação/Indústria Mineira de Joias/Entidades/Fiscal/Cupom/AdaptadorVarejo.cs
@@ -46,7 +46,7 @@ namespace Entidades.Fiscal.Cupom
             return new SaídaItemFiscal(AdaptarReferência(detalhe),
                                 detalhe.Descricao.Trim(),
                                 null,
-                                TipoUnidadeInterpretação.Interpretar(detalhe.Unidade),
+                                (int) TipoUnidadeInterpretação.Interpretar(detalhe.Unidade),
                                 detalhe.Quantidade,
                                 detalhe.ValorUnitario,
                                 detalhe.ValorTotalLiquido);
