+++ Programação/Indústria Mineira de Joias/Entidades/Fiscal/NotaFiscalEletronica/AdaptadorAtacadoEntrada.cs
@@ -33,7 +33,7 @@ namespace Entidades.Fiscal.NotaFiscalEletronica
                 itens.Add(new EntradaItemFiscal(parser.ObterReferência(x),
                     parser.ObterDescrição(x),
                     parser.ObterCFOP(x),
-                    parser.ObterTipoUnidade(x),
+                    (int) parser.ObterTipoUnidade(x),
                     parser.ObterQuantidadeItens(x),
                     parser.ObterValorUnitario(x),
                     parser.ObterValor(x)
